move "new field with section" tests to discrete module
